Update #244, #255
